refactoring (added movie class and swagger class)
